refactor: organize file structure
