+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
-import 'package:strecording/welcome_screen.dart';
-import 'package:strecording/setting_page.dart';
+import 'package:strecording/pages/welcome_screen.dart';
+import 'package:strecording/pages/setting_page.dart';
 
 void main() {
   KakaoSdk.init(nativeAppKey: 'a8a1d140843f9fde9e271aaba4a1412b');
