+++ lib/pages/login_page.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
-import 'package:strecording/login_platform.dart';
+import 'package:strecording/utilities/login_platform.dart';
 import 'package:google_sign_in/google_sign_in.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart';
 import 'package:flutter_facebook_auth/flutter_facebook_auth.dart';
